==> tutorials/4-collect-data.ipynb <==
## 4. Collect compliance jobs

### The final dataset has three components:
* listings with compliance keywords in title
* listings without compliance keywords in title but labeled positive by the BERT classifier
* listings with diversity keywords in title

import pandas as pd

df1 = pd.read_csv('data_for_filtering/job_with_compliance_keywords.csv')
compliance_keyword_df = df1[df1['label'] == 1]

df2 = pd.read_csv('data_for_filtering/negative_identifier_v3_07262022.csv')
compliance_model_df = df2[df2['negative_identifier_label'] == 1]

df3 = pd.read_csv('job_info_data/diversity_job_info_7212022.csv')

selected_columns = ['job_url', 'lnks_job_title', 'lnks_company', 'lnks_company_url', 'lnks_company_location', 'lnks_job_description']
final_df = pd.concat([compliance_keyword_df[selected_columns], df3[selected_columns], compliance_model_df[selected_columns]])
final_df.to_csv('compiled_dataset/compliance_jobs_june&july_2022.csv', index=False)

sampled_df = final_df.sample(n=300)
sampled_df.to_csv('compiled_dataset/compliance_jobs_june&july_2022_sampled_v1.csv', index=False)
